@@ -8,57 +8,51 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "@mapbox/geojsonhint": "^2.1.0",
-    "@turf/flatten": "^5.1.5",
-    "@turf/random": "^6.0.2",
-    "@turf/rewind": "^5.1.5",
-    "axios": "^0.19.0",
-    "file-saver": "^1.3.8",
-    "leaflet": "^1.4.0",
+    "core-js": "^3.8.3",
+    "vue": "^3.2.13",
+
+    "@turf/flatten": "^6.5.0",
+    "@turf/random": "^6.5.0",
+    "@turf/rewind": "^6.5.0",
+    "axios": "^1.2.4",
+    "file-saver": "^2.0.5",
+    "leaflet": "^1.9.3",
     "leaflet-draw": "^1.0.4",
-    "leaflet-simplestyle": "^1.0.1",
-    "leaflet-svg-shape-markers": "^1.2.0",
-    "shp-write": "^0.3.2",
-    "topojson-server": "^3.0.0",
-    "view-design": "^4.0.2",
-    "vue": "^2.5.22",
-    "vue-authenticate": "^1.4.1",
-    "vue-axios": "^2.1.5",
-    "vue-codemirror": "^4.0.6",
-    "vuex": "^3.0.1",
+    "leaflet-simplestyle": "^1.2.4",
+    "leaflet-svg-shape-markers": "^1.4.0",
+    "topojson-server": "^3.0.1",
+    "vue-authenticate": "^1.3.4",
+    "vue-axios": "^3.5.2",
+    "vue-codemirror": "^6.1.1",
+    "vuex": "^4.1.0",
     "wellknown": "^0.5.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^3.3.0",
-    "@vue/cli-plugin-eslint": "^3.3.0",
-    "@vue/cli-service": "^3.3.0",
-    "@vue/eslint-config-standard": "^4.0.0",
-    "babel-eslint": "^10.0.1",
-    "eslint": "^5.8.0",
-    "eslint-plugin-vue": "^5.0.0",
-    "sass": "^1.26.3",
-    "sass-loader": "^8.0.2",
-    "vue-template-compiler": "^2.5.22"
-  },
-  "babel": {
-    "presets": [
-      "@vue/app"
-    ]
+    "@babel/core": "^7.12.16",
+    "@babel/eslint-parser": "^7.12.16",
+    "@vue/cli-plugin-babel": "~5.0.0",
+    "@vue/cli-plugin-eslint": "~5.0.0",
+    "@vue/cli-service": "~5.0.0",
+    "eslint": "^7.32.0",
+    "eslint-plugin-vue": "^8.0.3",
+
+    "sass": "^1.57.1",
+    "sass-loader": "^13.2.0",
+    "vue-template-compiler": "^2.7.14"
   },
   "eslintConfig": {
     "root": true,
-    "extends": [
-      "plugin:vue/essential",
-      "eslint:recommended",
-      "@vue/standard"
-    ],
     "env": {
       "node": true
     },
-    "rules": {},
+    "extends": [
+      "plugin:vue/vue3-essential",
+      "eslint:recommended"
+    ],
     "parserOptions": {
-      "parser": "babel-eslint"
-    }
+      "parser": "@babel/eslint-parser"
+    },
+    "rules": {}
   },
   "postcss": {
     "plugins": {
@@ -68,6 +62,7 @@
   "browserslist": [
     "> 1%",
     "last 2 versions",
-    "not ie <= 8"
+    "not dead",
+    "not ie 11"
   ]
 }
