@@ -1,12 +1,4 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
-import ViewUI from 'view-design';
-import 'view-design/dist/styles/iview.css';
 
-Vue.use(ViewUI)
-
-Vue.config.productionTip = false
-
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
+createApp(App).mount('#app')
