@@ -1,7 +1,4 @@
-/* eslint-disable */
-module.exports = {
-  publicPath: process.env.NODE_ENV === 'production'
-    ? '/geojson-editor/'
-    : '/',
-  outputDir: 'docs'
-}
+const { defineConfig } = require('@vue/cli-service')
+module.exports = defineConfig({
+  transpileDependencies: true
+})
